write the central directory forward
Currently, the files contents are in lexicographic order, but the
filenames in the central directory are in reverse-lexicographic
order.

I found this because MS word thought .docx files created by this
library are suspicious (it prompts the user, but can read the file
just fine if you proceed), and I thought that the odd ordering could
trigger heuristics for "this looks like a zip-bomb".

I have lost track of whether this helps or not, but this is better
regardless.