support writing data descriptors (closes #4)
This is lightly tested, but at least a docx written by LibreOffice,
transformed through Zipc.{of,to}_binary_string, and then given
to MS Word works, unlike before.

AFAICT, the read size is fine because zipc ignores the extra data
(since all this data descriptor is already specified in the central
directory anyway).